Abolish sorting by commit timestamp in the emit stage.
Instead, we now use an algorithm something like a mutant toposort that
tries to pull each commit as far back in the repo is it can without
violating either time or topological ordering.

The objective, as before, is to emit a repository that has ordering as
though it had been dumped by got-fast-export (with the constraint that
this ordering is indeterminte in some edge cases, e.g. if there are
timestamp collisions).

This should make the 'Child commit emitted before parent exists' error
impossible; however, the check has bewen left in place.  All existing
tests pass, but I have yet to verify that the suite contains a sharp
test for the new logic.

Formally the new code has the same O(n**2) worst case as the quicksort
it replaced, but its average case is probably rather worse than O(n
log n). It's a bit simpler, though.