Address Gitlab error #61: syntax error on token @
The problem is an expand field that looks like "@b@" rather than "b".
I'm not actually sure which formn is correct, abd if both forms appear
in the wild we need to accept both.  So we add an alternation to the
YACC grammar to handle this.

Also add the master includeed in the bug report as the test altexpand,v.